+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 0.1
+## 0.2.0
+- Add the scripts for automatically updating NPM dependencies
+  - `autoup_minor` - Updates all dependencies to the latest minor version or patch version.
+  - `autoup` - Updates all dependencies to the latest version.
+
+## 0.1.0
 - Add the ability to automatically switch node versions based on the project's...
   - `.node-version` file
   - `.nvmrc` file